+++ src/main/java/blackjack/domain/game/Outcome.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
