+++ src/main/java/blackjack/domain/role/Dealer.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.role;
 
+import blackjack.domain.card.Card;
 import java.util.List;
 import java.util.function.Supplier;
