+++ src/main/java/blackjack/domain/role/Player.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.role;
 
+import blackjack.domain.card.Card;
 import java.util.List;
 
 public class Player extends Role {
