+++ src/main/java/blackjack/dto/DealerTableDto.java
@@ -1,7 +1,7 @@
 package blackjack.dto;
 
-import blackjack.domain.Card;
-import blackjack.domain.Role;
+import blackjack.domain.card.Card;
+import blackjack.domain.role.Role;
 import java.util.List;
 import java.util.stream.Collectors;
