+++ src/main/java/blackjack/dto/PlayerTableDto.java
@@ -3,8 +3,8 @@ package blackjack.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import blackjack.domain.Card;
-import blackjack.domain.Role;
+import blackjack.domain.card.Card;
+import blackjack.domain.role.Role;
 
 public class PlayerTableDto {
