+++ src/main/java/blackjack/service/BlackJackService.java
@@ -1,13 +1,13 @@
 package blackjack.service;
 
-import blackjack.domain.Compete;
-import blackjack.domain.Dealer;
-import blackjack.domain.Deck;
-import blackjack.domain.Hand;
-import blackjack.domain.Player;
-import blackjack.domain.PlayerTurns;
-import blackjack.domain.RedrawChoice;
-import blackjack.domain.Role;
+import blackjack.domain.game.Compete;
+import blackjack.domain.role.Dealer;
+import blackjack.domain.game.Deck;
+import blackjack.domain.role.Hand;
+import blackjack.domain.role.Player;
+import blackjack.domain.role.PlayerTurns;
+import blackjack.domain.game.RedrawChoice;
+import blackjack.domain.role.Role;
 import blackjack.dto.DealerTableDto;
 import blackjack.dto.DealerTurnDto;
 import blackjack.dto.FinalResultDto;
