+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,11 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import static org.assertj.core.api.AssertionsForClassTypes.*;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import blackjack.factory.CardMockFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import blackjack.domain.factory.CardMockFactory;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
