+++ src/test/java/blackjack/domain/card/DenominationTest.java
@@ -1,6 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import static org.assertj.core.api.AssertionsForClassTypes.*;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
