+++ src/test/java/blackjack/domain/role/DealerTest.java
@@ -1,18 +1,17 @@
-package blackjack.domain;
+package blackjack.domain.role;
 
-import static org.assertj.core.api.AssertionsForClassTypes.*;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import blackjack.domain.card.Card;
+import blackjack.factory.CardMockFactory;
+import blackjack.util.CreateHand;
 import java.util.List;
 import java.util.stream.Stream;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-
-import blackjack.domain.factory.CardMockFactory;
-import blackjack.domain.util.CreateHand;
 
 @DisplayName("Dealer 테스트")
 class DealerTest {
