+++ src/test/java/blackjack/domain/role/HandTest.java
@@ -1,9 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.role;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import blackjack.domain.factory.CardMockFactory;
-import blackjack.domain.util.CreateHand;
+import blackjack.domain.card.Card;
+import blackjack.factory.CardMockFactory;
+import blackjack.util.CreateHand;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
