+++ src/test/java/blackjack/factory/CardMockFactory.java
@@ -1,13 +1,13 @@
-package blackjack.domain.factory;
+package blackjack.factory;
 
 import java.util.Arrays;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import blackjack.domain.Card;
-import blackjack.domain.Denomination;
-import blackjack.domain.Suit;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.Suit;
 
 public class CardMockFactory {
 
@@ -15,8 +15,8 @@ public class CardMockFactory {
 
 	static {
 		cards = Arrays.stream(Denomination.values())
-			.map(denomination -> new Card(Suit.CLOVER, denomination))
-			.collect(Collectors.toMap(Card::getInformation, Function.identity()));
+				.map(denomination -> new Card(Suit.CLOVER, denomination))
+				.collect(Collectors.toMap(Card::getInformation, Function.identity()));
 	}
 
 	public static Card of(final String information) {
